Conflicts:
	Assets/slua_src/quaternion.cpp
	Assets/slua_src/slua.cpp
	Assets/slua_src/slua.hpp

+++ Assets/slua_src/vector3.hpp
@@ -23,9 +23,6 @@
 #ifndef SLUA_VECTOR3_HEADER_
 #define SLUA_VECTOR3_HEADER_
 
-#include <string>
-#include <algorithm>
-
 struct Vector3
 {
 
@@ -66,7 +63,7 @@ struct Vector3
 	// field
 	float magnitude() const
 	{
-		return std::sqrtf(x*x + y*y + z*z);
+		return std::sqrt(x*x + y*y + z*z);
 	}
 	Vector3 normalized() const;
 	float sqrMagnitude() const{
